The Point Evaluation precompile (eip-4844) benchmarks require Ckzg being initialized before run

The performance is good in comparison with other implementations ( https://github.com/imapp-pl/benchmarking/tree/precompiles_benchmark/shanghai#client-comparison )

```markdown
|   Method |            Input |       Mean |   Error |  StdDev | Ratio | Alloc Ratio |
|--------- |----------------- |-----------:|--------:|--------:|------:|------------:|
| Baseline |      fuzzcorp-33 |   109.7 us | 1.23 us | 2.93 us |  1.00 |        1.00 |
|          |                  |            |         |         |       |             |
| Baseline |      fuzzcorp-95 |   109.5 us | 0.77 us | 1.17 us |  1.00 |        1.00 |
|          |                  |            |         |         |       |             |
| Baseline | pointEvaluation1 | 1,104.5 us | 9.74 us | 8.13 us |  1.00 |        1.00 |
```
## Changes

- Add initialization
- Update library

## Types of changes

#### What types of changes does your code introduce?

- [x] Bugfix (a non-breaking change that fixes an issue)
- [ ] New feature (a non-breaking change that adds functionality)
- [ ] Breaking change (a change that causes existing functionality not to work as expected)
- [ ] Optimization
- [ ] Refactoring
- [ ] Documentation update
- [ ] Build-related changes
- [ ] Other: _Description_

## Testing

#### Requires testing

- [ ] Yes
- [x] No

#### If yes, did you write tests?

- [ ] Yes
- [x] No